+++ tests/test_query/aggregation/software_version/assert_software_version_aggregate.py
