+++ tests/test_query/aggregation/software_version/software_version_aggregate_fragment.py
